% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_print.R
\name{browsable}
\alias{browsable}
\alias{is.browsable}
\title{Make an HTML object browsable}
\usage{
browsable(x, value = TRUE)

is.browsable(x)
}
\arguments{
\item{x}{The object to make browsable or not.}

\item{value}{Whether the object should be considered browsable.}
}
\value{
`browsable` returns `x` with an extra attribute to indicate
  that the value is browsable.

`is.browsable` returns `TRUE` if the value is browsable, or
  `FALSE` if not.
}
\description{
By default, HTML objects display their HTML markup at the console when
printed. `browsable` can be used to make specific objects render as HTML
by default when printed at the console.
}
\details{
You can override the default browsability of an HTML object by explicitly
passing `browse = TRUE` (or `FALSE`) to the `print` function.
}
