% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{defaultPngDevice}
\alias{defaultPngDevice}
\title{Determine the best PNG device for your system}
\usage{
defaultPngDevice()
}
\value{
A graphics device function.
}
\description{
Returns the best PNG-based graphics device for your system, in the opinion of
the `htmltools` maintainers. On Mac,
[grDevices::png()] is used; on all other
platforms, either [ragg::agg_png()] or
[Cairo::CairoPNG()] are used if their packages
are installed. Otherwise, [grDevices::png()] is
used.
}
