% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{htlapply}
\alias{htlapply}
\title{lapply with progress bar}
\usage{
htlapply(X, FUN, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{expression} object.
Other objects (including classed objects) will be coerced by
\code{base::as.list}.}

\item{FUN}{the function to be applied to each element of X:
see ‘Details’. In the case of functions like +, %*%, the
function name must be backquoted or quoted.}

\item{...}{optional arguments to \code{FUN}.}
}
\description{
lapply with progress bar
}
