% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{plotTag}
\alias{plotTag}
\title{Capture a plot as a self-contained `<img>` tag}
\usage{
plotTag(
  expr,
  alt,
  device = defaultPngDevice(),
  width = 400,
  height = 400,
  pixelratio = 2,
  mimeType = "image/png",
  deviceArgs = list(),
  attribs = list(),
  suppressSize = c("none", "x", "y", "xy")
)
}
\arguments{
\item{expr}{A plotting expression that generates a plot (or yields an object
that generates a plot when printed, like a ggplot2).}

\item{alt}{A single-element character vector that contains a text description
of the image. This is used by accessibility tools, such as screen readers
for vision impaired users.}

\item{device}{A graphics device function; by default, this will be either
[grDevices::png()], [ragg::agg_png()], or [Cairo::CairoPNG()], depending on
your system and configuration. See [defaultPngDevice()].}

\item{width, height}{The width/height that the generated tag should be
displayed at, in logical (browser) pixels.}

\item{pixelratio}{Indicates the ratio between physical and logical units of
length. For PNGs that may be displayed on high-DPI screens, use `2`;
for graphics devices that express width/height in inches (like
[grDevices::svg()], try `1/72` or `1/96`.}

\item{mimeType}{The MIME type associated with the `device`. Examples are
`image/png`, `image/tiff`, `image/svg+xml`.}

\item{deviceArgs}{A list of additional arguments that should be included when
the `device` function is invoked.}

\item{attribs}{A list of additional attributes that should be included on the
generated `<img>` (e.g. `id`, `class`).}

\item{suppressSize}{By default, `plotTag` will include a `style`
attribute with `width` and `height` properties specified in
pixels. If you'd rather specify the image size using other methods (like
responsive CSS rules) you can use this argument to suppress width
(`"x"`), height (`"y"`), or both (`"xy"`) properties.}
}
\value{
A [browsable()] HTML `<img>` tag object. Print it at
  the console to preview, or call [as.character()] on it to view the HTML
  source.
}
\description{
Capture a plot as a self-contained `<img>` tag
}
\examples{
img <- plotTag({
  plot(cars)
}, "A plot of the 'cars' dataset", width = 375, height = 275)

if (interactive()) img

if (interactive() && capabilities("cairo")) {
  plotTag(
    plot(pressure), "A plot of the 'pressure' dataset",
    device = grDevices::svg, width = 375, height = 275, pixelratio = 1/72,
    mimeType = "image/svg+xml"
  )
}

}
\seealso{
[capturePlot()] saves plots as an image file.
}
