% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{renderTags}
\alias{renderTags}
\alias{doRenderTags}
\title{Render tags into HTML}
\usage{
renderTags(x, singletons = character(0), indent = 0)

doRenderTags(x, indent = 0)
}
\arguments{
\item{x}{Tag object(s) to render}

\item{singletons}{A list of [singleton] signatures to consider already
rendered; any matching singletons will be dropped instead of rendered.
(This is useful (only?) for incremental rendering.)}

\item{indent}{Initial indent level, or `FALSE` if no indentation should
be used.}
}
\value{
`renderTags` returns a list with the following variables:
  * `head`: An [HTML()] string that should be included in `<head>`.
  * `singletons`: Character vector of singleton signatures that are
  known after rendering.
  * `dependencies`: A list of [resolved][resolveDependencies] [htmlDependency()] objects.
  * `html`: An [HTML()] string that represents the main HTML that was rendered.

`doRenderTags` returns a simple [HTML()] string.
}
\description{
Renders tags (and objects that can be converted into tags using
[as.tags()]) into HTML. (Generally intended to be called from web
framework libraries, not directly by most users--see
[print.html()] for higher level rendering.)
}
\details{
`doRenderTags` is intended for very low-level use; it ignores
  render hooks, singletons, head, and dependency handling, and simply renders the given tag
  objects as HTML. Please use `renderTags()` if `x` has not already handled its dependencies
  and render hooks.
}
