% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tagAppendAttributes}
\alias{tagAppendAttributes}
\alias{tagHasAttribute}
\alias{tagGetAttribute}
\title{Append tag attributes}
\usage{
tagAppendAttributes(tag, ..., .cssSelector = NULL)

tagHasAttribute(tag, attr)

tagGetAttribute(tag, attr)
}
\arguments{
\item{tag}{a [tag] object.}

\item{...}{a collection of attributes.}

\item{.cssSelector}{A character string containing a [CSS
selector](https://developer.mozilla.org/en-US/docs/Learn/CSS/Building_blocks/Selectors)
for targeting particular (inner) tags of interest. At the moment, only a
combination of
[type](https://www.w3.org/TR/CSS22/selector.html#type-selectors) (e.g,
`div`), [class](https://www.w3.org/TR/CSS22/selector.html#class-html)
(e.g., `.my-class`),
[id](https://www.w3.org/TR/CSS22/selector.html#id-selectors) (e.g.,
`#myID`), and
[universal](https://www.w3.org/TR/CSS22/selector.html#universal-selector)
(`*`) selectors within a given [simple
selector](https://www.w3.org/TR/CSS22/selector.html#selector-syntax) is
supported. Note, if `.cssSelector` is used, the returned tags will have
their `$children` fields flattened to a single `list()` via [`tagQuery()`].}

\item{attr}{The name of an attribute.}
}
\description{
Append (`tagAppendAttributes()`), check existence (`tagHasAttribute()`),
and obtain the value (`tagGetAttribute()`) of HTML attribute(s).
}
\examples{
html <- div(a())
tagAppendAttributes(html, class = "foo")
tagAppendAttributes(html, .cssSelector = "a", class = "bar")

tagHasAttribute(div(foo = "bar"), "foo")
tagGetAttribute(div(foo = "bar"), "foo")

}
\seealso{
[tagAppendChildren()], [tagQuery()]
}
